Fixed thread stopping bug on gui exit
